+++ src/mobi/hsz/idea/gitignore/hideignoredfiles/HideIgnoredFilesTreeStructureProvider.java
@@ -1,6 +1,7 @@
 package mobi.hsz.idea.gitignore.hideignoredfiles;
 
-import com.intellij.ide.projectView.*;
+import com.intellij.ide.projectView.TreeStructureProvider;
+import com.intellij.ide.projectView.ViewSettings;
 import com.intellij.ide.projectView.impl.nodes.PsiDirectoryNode;
 import com.intellij.ide.projectView.impl.nodes.PsiFileNode;
 import com.intellij.ide.util.treeView.AbstractTreeNode;
@@ -9,10 +10,19 @@ import com.intellij.openapi.project.ProjectManager;
 import com.intellij.openapi.vfs.VirtualFile;
 import mobi.hsz.idea.gitignore.IgnoreManager;
 import mobi.hsz.idea.gitignore.settings.IgnoreSettings;
-import org.jetbrains.annotations.*;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
 
+/**
+ * Extension for the Project Tree View that provides it with the ability to hide from
+ * it files or folders that had been marked as ignored
+ *
+ * @author Maximiliano Najle <maximilianonajle@gmail.com>
+ * @since 1.7
+ */
 public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvider {
 
     private final Project project = ProjectManager.getInstance().getOpenProjects()[0];
@@ -24,7 +34,7 @@ public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvi
                                                @NotNull Collection<AbstractTreeNode> children,
                                                ViewSettings settings) {
 
-        if (! IgnoreSettings.getInstance().shouldHideIgnoredFilesOnProjectView()) {
+        if (!IgnoreSettings.getInstance().shouldHideIgnoredFilesOnProjectView()) {
             return children;
         }
 
@@ -47,13 +57,11 @@ public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvi
         try {
             VirtualFile file = ((PsiFileNode) node).getVirtualFile();
             return ignoreManager.isFileIgnored(file);
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             try {
                 VirtualFile file = ((PsiDirectoryNode) node).getVirtualFile();
                 return ignoreManager.isFileIgnored(file);
-            }
-            catch (Exception ex) {
+            } catch (Exception ex) {
                 return false;
             }
         }